Add mascot
This is a mascot image I originally created for the Rust Secure Code WG
but they decided to go with a Ferris variant instead.

It's a hermit crab made of bits and pieces of Ferris's vector graphics
which seems apropos. I nicknamed him "Hermes" but feel free to choose a
different name if you like.

This commit includes a raster PNG and SVG versions of the image
(I made it in Illustrator and can provide a .ai file if you like)